// src/lib/client/model/getFormTemplateRes.ts
/**
 * Generated by orval v6.21.0 🍺
 * Do not edit manually.
 * Nobstacle API
 * OpenAPI spec version: 1.0.0
 */
import type { FormValuesRes } from './formValuesRes';

export interface GetFormTemplateRes {
  id: number;
  langCode: string[];
  order: number;
  tag: string;
  templateId: number;
  values: FormValuesRes;
}